refactor: Line 역 목록 조회 기능 LineSections 도메인으로 위임 - LineSections 도메인 추출
