+++ src/main/java/nextstep/subway/line/domain/LineSections.java
@@ -1,0 +1,74 @@
+package nextstep.subway.line.domain;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+
+import nextstep.subway.station.domain.Station;
+
+@Embeddable
+public class LineSections {
+
+    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
+    private List<Section> lineSections = new ArrayList<>();
+
+    public LineSections() {
+    }
+
+    public LineSections(List<Section> lineSections) {
+        this.lineSections = lineSections;
+    }
+
+    public List<Station> toStation() {
+        if (lineSections.isEmpty()) {
+            return Arrays.asList();
+        }
+
+        List<Station> stations = new ArrayList<>();
+        Station downStation = findUpStation();
+        stations.add(downStation);
+
+        while (downStation != null) {
+            Station finalDownStation = downStation;
+            Optional<Section> nextLineStation = lineSections.stream()
+                .filter(it -> it.getUpStation() == finalDownStation)
+                .findFirst();
+            if (!nextLineStation.isPresent()) {
+                break;
+            }
+            downStation = nextLineStation.get().getDownStation();
+            stations.add(downStation);
+        }
+
+        return stations;
+    }
+
+    private Station findUpStation() {
+        Station downStation = lineSections.get(0).getUpStation();
+        while (downStation != null) {
+            Station finalDownStation = downStation;
+            Optional<Section> nextLineStation = lineSections.stream()
+                .filter(it -> it.getDownStation() == finalDownStation)
+                .findFirst();
+            if (!nextLineStation.isPresent()) {
+                break;
+            }
+            downStation = nextLineStation.get().getUpStation();
+        }
+
+        return downStation;
+    }
+
+    public void add(Section section) {
+        this.lineSections.add(section);
+    }
+
+    public List<Section> getLineSections() {
+        return lineSections;
+    }
+}
