+++ src/test/java/nextstep/subway/line/domain/LineSectionsTest.java
@@ -1,0 +1,5 @@
+package nextstep.subway.line.domain;
+
+public class LineSectionsTest {
+
+}
